fix and refactor sidebar layout
