+++ src/main/frontend/app/components/sidebars-layout/sidebar-layout-store.ts
