+++ src/main/frontend/app/routes/configurations/configuration-overview.tsx
@@ -14,7 +14,7 @@ import Search from '~/components/search/search'
 import SidebarLayout from '~/components/sidebars-layout/sidebar-layout'
 import SidebarHeader from '~/components/sidebars-layout/sidebar-header'
 import SidebarContentClose from '~/components/sidebars-layout/sidebar-content-close'
-import { SidebarSide, useSidebarStore } from '~/stores/sidebar-layout-store'
+import { SidebarSide, useSidebarStore } from '~/components/sidebars-layout/sidebar-layout-store'
 import NonCanvasComponentContext, { type NonCanvasComponentEditorState } from './non-canvas-component-context'
 import AdapterContext, { type AdapterEditorState } from './adapter-context'
 import NonCanvasComponentPalette from './non-canvas-component-palette'
